Add tests for GetUserIDsByExternalAccounts

+++ enterprise/cmd/frontend/db/perms_store_test.go
@@ -1574,17 +1574,23 @@ func testPermsStore_DatabaseDeadlocks(db *sql.DB) func(*testing.T) {
 	}
 }
 
+func cleanupUsersTable(t *testing.T, s *PermsStore) {
+	if t.Failed() {
+		return
+	}
+
+	q := `TRUNCATE TABLE users RESTART IDENTITY CASCADE;`
+	if err := s.execute(context.Background(), sqlf.Sprintf(q)); err != nil {
+		t.Fatal(err)
+	}
+}
+
 func testPermsStore_ListExternalAccounts(db *sql.DB) func(*testing.T) {
 	return func(t *testing.T) {
 		s := NewPermsStore(db, time.Now)
+		defer cleanupUsersTable(t, s)
 
 		ctx := context.Background()
-		defer func() {
-			q := `TRUNCATE TABLE users RESTART IDENTITY CASCADE;`
-			if err := s.execute(ctx, sqlf.Sprintf(q)); err != nil {
-				t.Fatal(err)
-			}
-		}()
 
 		// Set up test users and external accounts
 		extSQL := `
@@ -1592,12 +1598,12 @@ INSERT INTO user_external_accounts(user_id, service_type, service_id, account_id
 	VALUES(%s, %s, %s, %s, %s, %s, %s)
 `
 		qs := []*sqlf.Query{
-			sqlf.Sprintf(`INSERT INTO users(username) VALUES('alice')`), // Always has ID=1
-			sqlf.Sprintf(`INSERT INTO users(username) VALUES('bob')`),   // Always has ID=1
+			sqlf.Sprintf(`INSERT INTO users(username) VALUES('alice')`), // ID=1
+			sqlf.Sprintf(`INSERT INTO users(username) VALUES('bob')`),   // ID=2
 
-			sqlf.Sprintf(extSQL, 1, "gitlab", "https://gitlab.com/", "alice_gitlab", "alice_gitlab_client_id", clock(), clock()),
-			sqlf.Sprintf(extSQL, 1, "github", "https://github.com/", "alice_github", "alice_github_client_id", clock(), clock()),
-			sqlf.Sprintf(extSQL, 2, "gitlab", "https://gitlab.com/", "bob_gitlab", "bob_gitlab_client_id", clock(), clock()),
+			sqlf.Sprintf(extSQL, 1, "gitlab", "https://gitlab.com/", "alice_gitlab", "alice_gitlab_client_id", clock(), clock()), // ID=1
+			sqlf.Sprintf(extSQL, 1, "github", "https://github.com/", "alice_github", "alice_github_client_id", clock(), clock()), // ID=2
+			sqlf.Sprintf(extSQL, 2, "gitlab", "https://gitlab.com/", "bob_gitlab", "bob_gitlab_client_id", clock(), clock()),     // ID=3
 		}
 		for _, q := range qs {
 			if err := s.execute(ctx, q); err != nil {
@@ -1610,10 +1616,6 @@ INSERT INTO user_external_accounts(user_id, service_type, service_id, account_id
 			accounts, err := s.ListExternalAccounts(ctx, 1)
 			if err != nil {
 				t.Fatal(err)
-			}
-
-			if len(accounts) != 2 {
-				t.Fatalf("len(accounts): want 2 but got %d", len(accounts))
 			}
 
 			expAccounts := []*extsvc.ExternalAccount{
@@ -1654,10 +1656,6 @@ INSERT INTO user_external_accounts(user_id, service_type, service_id, account_id
 				t.Fatal(err)
 			}
 
-			if len(accounts) != 1 {
-				t.Fatalf("len(accounts): want 1 but got %d", len(accounts))
-			}
-
 			expAccounts := []*extsvc.ExternalAccount{
 				{
 					ID:     3,
@@ -1675,6 +1673,56 @@ INSERT INTO user_external_accounts(user_id, service_type, service_id, account_id
 			if diff := cmp.Diff(expAccounts, accounts); diff != "" {
 				t.Fatalf(diff)
 			}
+		}
+	}
+}
+
+func testPermsStore_GetUserIDsByExternalAccounts(db *sql.DB) func(t *testing.T) {
+	return func(t *testing.T) {
+		s := NewPermsStore(db, time.Now)
+		defer cleanupUsersTable(t, s)
+
+		ctx := context.Background()
+
+		// Set up test users and external accounts
+		extSQL := `
+INSERT INTO user_external_accounts(user_id, service_type, service_id, account_id, client_id, created_at, updated_at)
+	VALUES(%s, %s, %s, %s, %s, %s, %s)
+`
+		qs := []*sqlf.Query{
+			sqlf.Sprintf(`INSERT INTO users(username) VALUES('alice')`), // ID=1
+			sqlf.Sprintf(`INSERT INTO users(username) VALUES('bob')`),   // ID=2
+			sqlf.Sprintf(`INSERT INTO users(username) VALUES('cindy')`), // ID=3
+
+			sqlf.Sprintf(extSQL, 1, "gitlab", "https://gitlab.com/", "alice_gitlab", "alice_gitlab_client_id", clock(), clock()), // ID=1
+			sqlf.Sprintf(extSQL, 1, "github", "https://github.com/", "alice_github", "alice_github_client_id", clock(), clock()), // ID=2
+			sqlf.Sprintf(extSQL, 2, "gitlab", "https://gitlab.com/", "bob_gitlab", "bob_gitlab_client_id", clock(), clock()),     // ID=3
+			sqlf.Sprintf(extSQL, 3, "gitlab", "https://gitlab.com/", "cindy_gitlab", "cindy_gitlab_client_id", clock(), clock()), // ID=4
+		}
+		for _, q := range qs {
+			if err := s.execute(ctx, q); err != nil {
+				t.Fatal(err)
+			}
+		}
+
+		accounts := &extsvc.ExternalAccounts{
+			ServiceType: "gitlab",
+			ServiceID:   "https://gitlab.com/",
+			AccountIDs:  []string{"alice_gitlab", "bob_gitlab", "david_gitlab"},
+		}
+		userIDs, err := s.GetUserIDsByExternalAccounts(ctx, accounts)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		if len(userIDs) != 2 {
+			t.Fatalf("len(userIDs): want 2 but got %v", userIDs)
+		}
+
+		if userIDs["alice_gitlab"] != 1 {
+			t.Fatalf(`userIDs["alice_gitlab"]: want 1 but got %d`, userIDs["alice_gitlab"])
+		} else if userIDs["bob_gitlab"] != 2 {
+			t.Fatalf(`userIDs["bob_gitlab"]: want 2 but got %d`, userIDs["bob_gitlab"])
 		}
 	}
 }
